requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
django-html-field==0.1.3
django-restframework==0.0.1
django-tinymce==3.4.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
Pillow==9.0.0
# python-magic==0.4.24
# python-magic-bin==0.4.14
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
python-decouple